Trait cfg::rule_container::RuleContainer

pub trait RuleContainer: SymbolSource {
    type History;
    fn retain<F>(&mut self, f: F) where F: FnMut(Self::Symbol, &[Self::Symbol], &Self::History) -> bool;
    fn add_rule(&mut self, lhs: Self::Symbol, rhs: &[Self::Symbol], history: Self::History);
}

Trait for rule and symbol containers.

Associated Types

type History

The type of history carried with the rule.

Required Methods

fn retain<F>(&mut self, f: F) where F: FnMut(Self::Symbol, &[Self::Symbol], &Self::History) -> bool

Retains only the rules specified by the predicate.

In other words, removes all rules rule for which f(&rule) returns false.

fn add_rule(&mut self, lhs: Self::Symbol, rhs: &[Self::Symbol], history: Self::History)

Inserts a rule with lhs and rhs on its LHS and RHS. The rule carries history.