Module cfg::sequence

Sequences are similar to regex repetitions with numbering.

Modules

builder
Sequence rules can be built with the builder pattern.

destination
Sequence destination.

rewrite
Rewrites sequence rules into production rules.

Structs

Sequence
Sequence rule representation.

Enums

Separator
The separator symbol and mode of separation in a sequence, or Null for no separation.