Struct cfg::sequence_builder::SequenceRuleBuilder

pub struct SequenceRuleBuilder<H, D, S> where S: GrammarSymbol {
    // some fields omitted
}

Sequence rule builder.

Methods

impl<H, D, S> SequenceRuleBuilder<H, D, S> where D: SequenceDestination<H, Symbol=S>, H: RewriteSequence, S: GrammarSymbol

fn new(destination: D) -> Self

Creates a sequence rule builder.

fn sequence(self, lhs: S) -> Self

Starts building a sequence rule.

fn separator(self, sep: Separator<S>) -> Self

Assigns the separator symbol and mode of separation.

fn intersperse(self, sym: S) -> Self

Sets proper separation with the given separator symbol.

fn history(self, history: H) -> Self

Assigns the rule history, which is used on the next call to rhs, or overwritten by a call to rhs_with_history.

fn inclusive(self, start: u32, end: Option<u32>) -> Self

Assigns the inclusive range of the number of repetitions.

fn rhs<T>(self, rhs: S, range: T) -> Self where T: RangeArgument<u32>, H: Default

Adds a sequence rule to the grammar.

fn rhs_with_history(self, rhs: S, history: H) -> Self

Adds a sequence rule to the grammar.