Module cfg::symbol

Symbol types can be used to parameterize grammars.

Structs

ConsecutiveSymbols
A source of numeric symbols.

Nonterminals
Iterator for generating nonterminal symbols.

NumericSymbol
A numeric symbol type.

Terminals
Iterator for generating terminal symbols.

Traits

GrammarSymbol
A type that can represent symbols in a context-free grammar. Symbols are distinguished by their IDs.

SymbolContainer
Trait used to generate terminal symbols.

SymbolSource
A source of symbols.

TerminalSymbolSet
A source of symbols that tracks whether a symbol is terminal or nonterminal.