Trait cfg::ContextFreeRef

pub trait ContextFreeRef<'a>: Deref where Self::Target: ContextFree {
    type RuleRef: GrammarRule<History=Self::Target::History> + Copy + 'a;
    type Rules: Iterator<Item=Self::RuleRef>;
    fn rules(self) -> Self::Rules;
}

This trait is currently needed to make the associated Rules iterator generic over a lifetime parameter.

Associated Types

type RuleRef: GrammarRule<History=Self::Target::History> + Copy + 'a

Immutable reference to a rule.

type Rules: Iterator<Item=Self::RuleRef>

Iterator over immutable references to the grammar's rules.

Required Methods

fn rules(self) -> Self::Rules

Returns an iterator over immutable references to the grammar's rules.