Crate cfg

Library for manipulations on context-free grammars. Most transformations are abstracted over grammar representations.

Reexports

pub use symbol::SymbolSource;

Modules

cycles
Cycle detection and elimination.

history
Any data carried alongside a grammar rule can be its history. Rule histories may contain more than semantic actions.

precedence
Precedenced rules are built with the builder pattern.

prediction
Prediction for predictive parsers.

rule_builder
Grammar rules can be built with the builder pattern.

rule_container
Abstraction for collections of rules.

sequence
Sequences are similar to regex repetitions with numbering.

sequence_builder
Sequence rules can be built with the builder pattern.

symbol
Symbol types can be used to parameterize grammars.

usefulness
Analysis of rule usefulness.

Structs

BinarizedCfg
Representation for grammars where right-hand sides of all rules have at most two symbols.

Cfg
Basic representation of context-free grammars.

Traits

ContextFree
Trait for context-free grammars.

ContextFreeMut
Allows access to a ContextFreeRef through mutable references.

ContextFreeRef
This trait is currently needed to make the associated Rules iterator generic over a lifetime parameter.

GrammarRule
Trait for rules of a context-free grammar.