Crate cfg

cfg defines conditional compiling options, cfg attibute parser and evaluator

Structs

CfgDiff

CfgOptions
Configuration options used for conditional compilation on items with cfg attributes. We have two kind of options in different namespaces: atomic options like unix, and key-value options like target_arch="x86".

DnfExpr
A #[cfg] directive in Disjunctive Normal Form (DNF).

InactiveReason

Enums

CfgAtom
A simple configuration value passed in from the outside.

CfgExpr