Module cfg::history

Any data carried alongside a grammar rule can be its history. Rule histories may contain more than semantic actions.

Structs

NullHistory
A history which carries no data. All operations on NullHistory are no-op.

Enums

BinarizedRhsSubset
Used to inform which symbols on a rule's RHS are nullable, and will be eliminated.

Traits

Action
Trait for history types that may have semantic actions.

AssignPrecedence
Trait for history types that allow the rule to have its precedence assigned.

Binarize
Trait for history types that allow the rule to be binarized.

EliminateNulling
Trait for history types that allow the rule to have nulling symbols eliminated from the RHS.

RewriteSequence
Trait for history types that allow the sequence rule to be rewritten into grammar rules.