Struct cfg::BinarizedCfg

pub struct BinarizedCfg<H = NullHistory, Ss = ConsecutiveSymbols> where Ss: SymbolSource {
    // some fields omitted
}

Representation for grammars where right-hand sides of all rules have at most two symbols.

Methods

impl<H> BinarizedCfg<H>

fn new() -> BinarizedCfg<H>

Creates a BinarizedCfg.

impl<H, Ss> BinarizedCfg<H, Ss> where Ss: SymbolSource

fn with_sym_source(sym_source: Ss) -> BinarizedCfg<H, Ss>

Creates an empty BinarizedCfg with the given symbol source.

fn from_context_free<'a, G>(this: &'a G) -> BinarizedCfg<H, Ss> where G: ContextFree<History=H, Source=Ss>, &'a G: ContextFreeRef<'a, Target=G>, H: Binarize + Clone + 'static, Ss: Clone + SymbolSource<Symbol=G::Symbol>

Creates a BinarizedCfg by binarizing a context-free grammar.

fn sort(&mut self)

Sorts the rule array.

fn dedup(&mut self)

Removes consecutive duplicate rules.

impl<H, Ss> BinarizedCfg<H, Ss> where H: Binarize + Clone + EliminateNulling, Ss: SymbolSource + Clone

fn eliminate_nulling_rules(&mut self) -> BinarizedCfg<H, Ss>

Eliminates all rules of the form A ::= epsilon.

In other words, this splits off the set of nulling rules.

The language represented by the grammar is preserved, except for the possible lack of the empty string. Unproductive rules aren't preserved.